+++ packages/spacecat-shared-data-access/src/models/site/config.js
@@ -19,6 +19,9 @@ export const configSchema = Joi.object({
     invitedUserCount: Joi.number().integer().min(0),
   }),
   imports: Joi.array().items(Joi.object({ type: Joi.string() }).unknown(true)),
+  fetchConfig: Joi.object({
+    headers: Joi.object().pattern(Joi.string(), Joi.string()),
+  }).optional(),
   handlers: Joi.object().pattern(Joi.string(), Joi.object({
     mentions: Joi.object().pattern(Joi.string(), Joi.array().items(Joi.string())),
     excludedURLs: Joi.array().items(Joi.string()),
@@ -45,8 +48,7 @@ export const configSchema = Joi.object({
 
 export const DEFAULT_CONFIG = {
   slack: {},
-  handlers: {
-  },
+  handlers: {},
 };
 
 // Function to validate incoming configuration
@@ -77,6 +79,7 @@ export const Config = (data = {}) => {
   self.getIncludedURLs = (type) => state?.handlers?.[type]?.includedURLs;
   self.getGroupedURLs = (type) => state?.handlers?.[type]?.groupedURLs;
   self.getLatestMetrics = (type) => state?.handlers?.[type]?.latestMetrics;
+  self.getFetchConfig = () => state?.fetchConfig;
 
   self.updateSlackConfig = (channel, workspace, invitedUserCount) => {
     state.slack = {
@@ -129,6 +132,10 @@ export const Config = (data = {}) => {
     state.handlers[type].latestMetrics = latestMetrics;
   };
 
+  self.updateFetchConfig = (fetchConfig) => {
+    state.fetchConfig = fetchConfig;
+  };
+
   return Object.freeze(self);
 };
 
@@ -138,4 +145,5 @@ Config.toDynamoItem = (config) => ({
   slack: config.getSlackConfig(),
   handlers: config.getHandlers(),
   imports: config.getImports(),
+  fetchConfig: config.getFetchConfig(),
 });